flower/requirements.txt
alipay-sdk-python==3.7.796
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
cffi==2.0.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.1
Django==5.2
django-bootstrap3==25.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
pycryptodomex==3.23.0
pygame==2.6.1
PyMySQL==1.1.1
pyOpenSSL==25.3.0
python-alipay-sdk==3.3.0
python-dateutil==2.9.0.post0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
gunicorn>=20.1.0
